save current process. The eliminating of data error is gone and the only error is the inital context bug that save the context enters in a new one causing a page reload.
