+++ shift-scan/messages/en.json
@@ -55,7 +55,9 @@
         "Equipment-result": "Equipment Found:",
         "Equipment-notes-title": "Notes",
         "Next-btn": "Next",
-        "recentlyUsed" : "Recently Used"
+        "recentlyUsed" : "Recently Used",
+        "SearchedCodes" : "Current Results",
+        "noResults" : "No results found"
     },
 
     "Footer": {
@@ -63,7 +65,7 @@
     },
     
     "dashboard": {
-        "Banner": "End of the line pal!",
+        "Banner": "Working towards success!",
         "Name": "{firstName} {lastName}",
         "Date": "{date}",
         "lN1": "Copyright © 2024.Streamline Precision LLC. All rights reserved."
